Drop Hugging Face from training; export a half-precision zip instead
No Drive and no external accounts: training runs on the machine's own
disk with the same resumable checkpoints, and finishes by writing
intent-classifier.zip - the best model saved in half precision, about
half the size, with its tokenizer and label map. The server's classifier
loads it back as float32 so it runs correctly on CPU too.

Checked by exporting a model, unzipping it into a fresh directory and
loading it through the server's own classifier: loads as float32 and
predicts correctly. The notebook ends with a browser download of the zip
and the two commands to install it on the server.

+++ backend/app/ai/intent/classifier.py
@@ -38,7 +38,9 @@ def _load():
 
     device = registry.resolve_device()
     tokenizer = AutoTokenizer.from_pretrained(str(path))
-    model = AutoModelForSequenceClassification.from_pretrained(str(path))
+    # The exported model is stored in half precision to keep the download small;
+    # load it as float32 so it runs correctly on CPU as well as GPU.
+    model = AutoModelForSequenceClassification.from_pretrained(str(path), torch_dtype=torch.float32)
     model.to(device).eval()
 
     labels_file = path / "labels.json"
